+++ R/RcppMeCab-package.r
@@ -1,19 +1,17 @@
 #' RcppMeCab: Rcpp Wrapper for MeCab Library
 #'
-#' R package based on \code{Rcpp} for \code{MeCab}: Yet Another Part-of-Speech and
+#' R package based on 'Rcpp' for 'MeCab': Yet Another Part-of-Speech and
 #' Morphological Analyzer (\url{http://taku910.github.io/mecab/}). The purpose of
 #' this package is providing a seamless developing and analyzing environment for
 #' CJK texts. This package utilizes parallel programming for providing
 #' highly efficient text preprocessing \code{posParallel()} function.
-#' For installation, please refer to README.md file.
+#' For installation, please refer to \href{https://github.com/junhewk/RcppMeCab/blob/master/README.md}{README.md} file.
 #'
 #' @name RcppMeCab
 #' @docType package
 #' @author Junhewk Kim Taku Kudo
-#' @import Rcpp
-#' @importFrom RcppParallel RcppParallelLibs
 #' @details
-#' This package utilizes \code{MeCab} C API and \code{Rcpp} codes.
+#' This package utilizes 'MeCab' C API and 'Rcpp' codes.
 #'
 #' @references
 #' \itemize{